backend/requirements_stt.txt
openai-whisper
sounddevice
numpy<2.4
scipy
imageio-ffmpeg